+++ library-model/library-model-generator-integration-test/src/test/java/com/uber/nullaway/libmodel/LibraryModelIntegrationTest.java
@@ -2,7 +2,7 @@ package com.uber.nullaway.libmodel;
 
 import com.google.errorprone.CompilationTestHelper;
 import com.uber.nullaway.NullAway;
-import com.uber.nullaway.testhelper.NullAwayJSpecifyConfig;
+import com.uber.nullaway.generics.JSpecifyJavacConfig;
 import java.util.Arrays;
 import org.junit.Before;
 import org.junit.Rule;
@@ -133,7 +133,7 @@ public class LibraryModelIntegrationTest {
   public void libraryModelInnerClassNullableUpperBoundsTest() {
     compilationHelper
         .setArgs(
-            NullAwayJSpecifyConfig.withJSpecifyModeArgs(
+            JSpecifyJavacConfig.withJSpecifyModeArgs(
                 Arrays.asList(
                     "-d",
                     temporaryFolder.getRoot().getAbsolutePath(),
@@ -160,7 +160,7 @@ public class LibraryModelIntegrationTest {
   public void libraryModelNullableUpperBoundsWithoutJarInferTest() {
     compilationHelper
         .setArgs(
-            NullAwayJSpecifyConfig.withJSpecifyModeArgs(
+            JSpecifyJavacConfig.withJSpecifyModeArgs(
                 Arrays.asList(
                     "-d",
                     temporaryFolder.getRoot().getAbsolutePath(),
@@ -181,7 +181,7 @@ public class LibraryModelIntegrationTest {
   public void libraryModelDefaultParameterNullabilityTest() {
     compilationHelper
         .setArgs(
-            NullAwayJSpecifyConfig.withJSpecifyModeArgs(
+            JSpecifyJavacConfig.withJSpecifyModeArgs(
                 Arrays.asList(
                     "-d",
                     temporaryFolder.getRoot().getAbsolutePath(),
@@ -205,7 +205,7 @@ public class LibraryModelIntegrationTest {
   public void libraryModelParameterNullabilityTest() {
     compilationHelper
         .setArgs(
-            NullAwayJSpecifyConfig.withJSpecifyModeArgs(
+            JSpecifyJavacConfig.withJSpecifyModeArgs(
                 Arrays.asList(
                     "-d",
                     temporaryFolder.getRoot().getAbsolutePath(),
@@ -232,7 +232,7 @@ public class LibraryModelIntegrationTest {
   public void nullableArrayTest() {
     compilationHelper
         .setArgs(
-            NullAwayJSpecifyConfig.withJSpecifyModeArgs(
+            JSpecifyJavacConfig.withJSpecifyModeArgs(
                 Arrays.asList(
                     "-d",
                     temporaryFolder.getRoot().getAbsolutePath(),
@@ -258,7 +258,7 @@ public class LibraryModelIntegrationTest {
   public void genericParameterTest() {
     compilationHelper
         .setArgs(
-            NullAwayJSpecifyConfig.withJSpecifyModeArgs(
+            JSpecifyJavacConfig.withJSpecifyModeArgs(
                 Arrays.asList(
                     "-d",
                     temporaryFolder.getRoot().getAbsolutePath(),
