+++ nullaway/src/test/java/com/uber/nullaway/jspecify/GenericMethodTests.java
@@ -2,7 +2,7 @@ package com.uber.nullaway.jspecify;
 
 import com.google.errorprone.CompilationTestHelper;
 import com.uber.nullaway.NullAwayTestsBase;
-import com.uber.nullaway.testhelper.NullAwayJSpecifyConfig;
+import com.uber.nullaway.generics.JSpecifyJavacConfig;
 import java.util.Arrays;
 import org.junit.Ignore;
 import org.junit.Test;
@@ -1400,13 +1400,13 @@ public class GenericMethodTests extends NullAwayTestsBase {
 
   private CompilationTestHelper makeHelper() {
     return makeTestHelperWithArgs(
-        NullAwayJSpecifyConfig.withJSpecifyModeArgs(
+        JSpecifyJavacConfig.withJSpecifyModeArgs(
             Arrays.asList("-XepOpt:NullAway:AnnotatedPackages=com.uber")));
   }
 
   private CompilationTestHelper makeHelperWithInferenceFailureWarning() {
     return makeTestHelperWithArgs(
-        NullAwayJSpecifyConfig.withJSpecifyModeArgs(
+        JSpecifyJavacConfig.withJSpecifyModeArgs(
             Arrays.asList(
                 "-XepOpt:NullAway:AnnotatedPackages=com.uber",
                 "-XepOpt:NullAway:WarnOnGenericInferenceFailure=true")));
